# Conflicts:
#	psychopy/tests/test_visual/test_basevisual.py
#	psychopy/visual/target.py

+++ buildWinDistributions.ps1
@@ -1,5 +1,6 @@
 #! powershell
 
+param ($install_pp=1)
 # build simple distributions
 # python setup.py bdist_egg
 # python setup.py sdist --formats=zip
@@ -9,8 +10,8 @@
 # $pyPaths = @("C:\Python36\", "C:\Python36_64\")
 # $names = @("PsychoPy3", "PsychoPy3")
 # $archs = @("win32", "win64")
-$pyPaths = @("C:\Python38\", "C:\Python36_64\")
-$names = @("PsychoPy_py38", "PsychoPy_py36")
+$pyPaths = @("C:\Python38\")
+$names = @("PsychoPy")
 $archs = @("win64")
 
 # read from the version file
@@ -21,29 +22,31 @@ for ($i=0; $i -lt $pyPaths.Length; $i++) {
     [console]::beep(440,300); [console]::beep(880,300)
     # try to uninstall psychopy from site-packages
     # re-install the current version as editable/developer
-    Invoke-Expression ("& '{0}python.exe' -m pip install . --no-deps --force" -f $pyPaths[$i])
-    echo ("Installed current PsychoPy")
-    xcopy /I /Y psychopy\*.txt $pyPaths[$i]
+    if ($install_pp -eq 1) {
+        Invoke-Expression ("& '{0}python.exe' -m pip install . --no-deps --force" -f $pyPaths[$i])
+        echo ("Installed current PsychoPy")
+        xcopy /I /Y psychopy\*.txt $pyPaths[$i]
+    }
 
     # build the installer
     $thisPath = $pyPaths[$i]
     $thisName = $names[$i]
     $thisArch = $archs[$i]
-    $cmdStr = "makensis.exe /v2 /DPRODUCT_VERSION={0} /DPRODUCT_NAME={1} /DARCH={2} /DPYPATH={3} buildCompleteInstaller.nsi" -f $v, $thisName, $thisArch, $thisPath
+    $cmdStr = "makensis.exe /v3 /DPRODUCT_VERSION={0} /DPRODUCT_NAME={1} /DARCH={2} /DPYPATH={3} buildCompleteInstaller.nsi" -f $v, $thisName, $thisArch, $thisPath
     echo $cmdStr
     Invoke-Expression $cmdStr
     # "C:\Program Files\Caphyon\Advanced Installer 13.1\bin\x86\AdvancedInstaller.com" /rebuild PsychoPy_AdvancedInstallerProj.aip
 
-    echo 'moving files to ..\dist'
-
-    Invoke-Expression ("& '{0}python.exe' setup.py clean --all" -f $pyPaths[$i])  # clean up our build dir
-    # try to uninstall psychopy from site-packages
-    Invoke-Expression ("& '{0}python.exe' -m pip uninstall -y psychopy" -f $pyPaths[$i])
-    # re-install the current version as editable/developer
-    Invoke-Expression ("& '{0}python.exe' -m pip install -e . --no-deps" -f $pyPaths[$i])
-
+    if ($install_pp -eq 1) {
+        Invoke-Expression ("& '{0}python.exe' setup.py clean --all" -f $pyPaths[$i])  # clean up our build dir
+        # try to uninstall psychopy from site-packages
+        Invoke-Expression ("& '{0}python.exe' -m pip uninstall -y psychopy" -f $pyPaths[$i])
+        # re-install the current version as editable/developer
+        Invoke-Expression ("& '{0}python.exe' -m pip install -e . --no-deps" -f $pyPaths[$i])
+    }
 }
 
+echo 'moving files to ..\dist'
 Move-Item -Force "StandalonePsychoPy*.exe" "..\dist"
 Move-Item -Force dist\* "..\dist"
